+++ diesel_tests/tests/associations.rs
@@ -6,8 +6,8 @@ fn one_to_many_returns_query_source_for_association() {
     let connection = connection_with_sean_and_tess_in_users_table();
     setup_posts_table(&connection);
 
-    let sean: User = connection.find(users::table, 1).unwrap().unwrap();
-    let tess: User = connection.find(users::table, 2).unwrap().unwrap();
+    let sean: User = connection.find(users::table, 1).unwrap();
+    let tess: User = connection.find(users::table, 2).unwrap();
     let seans_posts: Vec<Post> = connection.insert(&posts::table, &vec![
         sean.new_post("Hello", None), sean.new_post("World", None),
     ]).unwrap().collect();